This is meant to replace views in the FastDEC module with regular ACSet accessing. I've also taken the liberty of cleaning up some strange logic.