@GeorgeR227 2024-06-04
A few benchmarks of note. Overall performance otherwise is preserved.

### Old:
Operator: Geometric Hodge
Variant: New Form-1, [0.935041 ms, 2.212816 MB]
Variant: Old Form-1, [53.808375 ms, 54.740096 MB]

Operator: Wedge Product
Variant: New Form-0, Form-1, [0.052583 ms, 0.308864 MB]
Variant: New Form-0, Form-2, [0.339291 ms, 0.415056 MB]
Variant: New Form-1, Form-1, [0.1625 ms, 0.35056 MB]
Variant: Old Form-0, Form-1, [8.164167 ms, 8.05008 MB]
Variant: Old Form-0, Form-2, [46.512188 ms, 83.6008 MB]
Variant: Old Form-1, Form-1, [50.183625 ms, 56.1576 MB]

### New:
Operator: Geometric Hodge
Variant: New Form-1, [0.9962295 ms, 1.8428 MB]

Operator: Wedge Product
Variant: New Form-0, Form-1, [0.053959 ms, 0.308864 MB]
Variant: New Form-0, Form-2, [0.257666 ms, 0.409744 MB]
Variant: New Form-1, Form-1, [0.154208 ms, 0.35056 MB]