@GeorgeR227 2024-06-04
```julia
Test Summary: | Pass  Total   Time
CUDA          |  139    139  41.7s
     Testing CombinatorialSpaces tests passed 

shell> git log -1
commit 67ad092ad6615e66994ce3956d7a65657fdf3684 (HEAD -> gr/clean-fast-dec, origin/gr/clean-fast-dec)
Author: GeorgeR227 <78235421+GeorgeR227@users.noreply.github.com>
Date:   Tue Jun 4 15:32:09 2024 -0400

    Added some doc strings.
```